Convert typography to TS
